# src/frontend/openapi.json
{
  "components": {
    "schemas": {
      "Body_report_milestone_api_loans__loan_id__milestones_post": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "photos": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Photos",
            "type": "array"
          },
          "stage": {
            "anyOf": [
              {
                "enum": [
                  "foundation",
                  "plinth",
                  "slab",
                  "brickwork_roof",
                  "finishing"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          }
        },
        "required": [
          "photos"
        ],
        "title": "Body_report_milestone_api_loans__loan_id__milestones_post",
        "type": "object"
      },
...
      "BuildProgressView": {
        "properties": {
          "borrower": {
            "title": "Borrower",
            "type": "string"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "disbursed": {
            "title": "Disbursed",
            "type": "number"
          },
          "last_verified_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified On"
          },
          "loan_id": {
            "title": "Loan Id",
            "type": "string"
          },
          "locality": {
            "title": "Locality",
            "type": "string"
          },
          "paused": {
            "title": "Paused",
            "type": "boolean"
          },
          "phases": {
            "items": {
              "$ref": "#/components/schemas/PhaseHistoryView"
            },
            "title": "Phases",
            "type": "array"
          },
          "plot_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Label"
          },
          "sanctioned": {
            "title": "Sanctioned",
            "type": "number"
          },
          "shortfall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shortfall"
          },
          "standing": {
            "items": {
              "$ref": "#/components/schemas/StandingRowView"
            },
            "title": "Standing",
            "type": "array"
          },
          "steps": {
            "items": {
              "type": "string"
            },
            "title": "Steps",
            "type": "array"
          },
          "tranches": {
            "items": {
              "$ref": "#/components/schemas/ProgressTrancheView"
            },
            "title": "Tranches",
            "type": "array"
          }
        },
        "required": [
          "loan_id",
          "borrower",
          "locality",
          "plot_label",
          "sanctioned",
          "disbursed",
          "current_stage",
          "last_verified_on",
          "paused",
          "tranches",
          "standing",
          "steps",
          "shortfall",
          "phases"
        ],
        "title": "BuildProgressView",
        "type": "object"
      },
...
      "DecisionRequest": {
        "properties": {
          "action": {
            "enum": [
              "RELEASE",
              "HOLD",
              "ESCALATE"
            ],
            "title": "Action",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "action"
        ],
        "title": "DecisionRequest",
        "type": "object"
      },
      "DecisionResponse": {
        "properties": {
          "action": {
            "enum": [
              "RELEASE",
              "HOLD",
              "ESCALATE"
            ],
            "title": "Action",
            "type": "string"
          },
          "decided_at": {
            "format": "date-time",
            "title": "Decided At",
            "type": "string"
          },
          "loan_id": {
            "title": "Loan Id",
            "type": "string"
          },
          "tranche": {
            "title": "Tranche",
            "type": "integer"
          }
        },
        "required": [
          "loan_id",
          "tranche",
          "action",
          "decided_at"
        ],
        "title": "DecisionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoanSummaryView": {
        "description": "The loan itself \u2014 what every owner and bank header needs to name it.\n\nDeliberately not screen-shaped: it carries the loan's own facts, and each\nscreen picks the subset it renders.",
        "properties": {
          "borrower": {
            "title": "Borrower",
            "type": "string"
          },
          "built_up_sqft": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Built Up Sqft"
          },
          "contractor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contractor"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "disbursed": {
            "title": "Disbursed",
            "type": "number"
          },
          "exposure": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exposure"
          },
          "gap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap"
          },
          "latest_rev": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Rev"
          },
          "loan_id": {
            "title": "Loan Id",
            "type": "string"
          },
          "locality": {
            "title": "Locality",
            "type": "string"
          },
          "plot_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Label"
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          },
          "sanctioned": {
            "title": "Sanctioned",
            "type": "number"
          },
          "tranche_count": {
            "title": "Tranche Count",
            "type": "integer"
          }
        },
        "required": [
          "loan_id",
          "borrower",
          "locality",
          "plot_label",
          "built_up_sqft",
          "sanctioned",
          "disbursed",
          "contractor",
          "latest_rev",
          "recommendation",
          "exposure",
          "gap",
          "current_stage",
          "tranche_count"
        ],
        "title": "LoanSummaryView",
        "type": "object"
      },
...
      "MilestoneResponse": {
        "properties": {
          "photos": {
            "title": "Photos",
            "type": "integer"
          },
          "tranche": {
            "title": "Tranche",
            "type": "integer"
          }
        },
        "required": [
          "tranche",
          "photos"
        ],
        "title": "MilestoneResponse",
        "type": "object"
      },
...
      "PhotoView": {
        "properties": {
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "chips": {
            "items": {
              "$ref": "#/components/schemas/EvidenceChipView"
            },
            "title": "Chips",
            "type": "array"
          },
          "slot_key": {
            "title": "Slot Key",
            "type": "string"
          }
        },
        "required": [
          "slot_key",
          "caption",
          "chips"
        ],
        "title": "PhotoView",
        "type": "object"
      },
...
      "QuestionsSentResponse": {
        "properties": {
          "sent": {
            "title": "Sent",
            "type": "integer"
          }
        },
        "required": [
          "sent"
        ],
        "title": "QuestionsSentResponse",
        "type": "object"
      },
...
      "TrancheDecisionView": {
        "properties": {
          "borrower": {
            "title": "Borrower",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "exposure": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exposure"
          },
          "exposure_undefined": {
            "title": "Exposure Undefined",
            "type": "boolean"
          },
          "loan_id": {
            "title": "Loan Id",
            "type": "string"
          },
          "locality": {
            "title": "Locality",
            "type": "string"
          },
          "math": {
            "items": {
              "$ref": "#/components/schemas/MathRowView"
            },
            "title": "Math",
            "type": "array"
          },
          "milestone": {
            "title": "Milestone",
            "type": "string"
          },
          "needs_human_review": {
            "title": "Needs Human Review",
            "type": "boolean"
          },
          "officer_view": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Officer View"
          },
          "owner_view": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner View"
          },
          "phases": {
            "items": {
              "$ref": "#/components/schemas/PhaseHistoryView"
            },
            "title": "Phases",
            "type": "array"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PhotoView"
            },
            "title": "Photos",
            "type": "array"
          },
          "recommendation": {
            "title": "Recommendation",
            "type": "string"
          },
          "recommendation_tone": {
            "enum": [
              "danger",
              "warn",
              "success",
              "neutral"
            ],
            "title": "Recommendation Tone",
            "type": "string"
          },
          "request_amount": {
            "title": "Request Amount",
            "type": "number"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageView"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tranche_number": {
            "title": "Tranche Number",
            "type": "integer"
          }
        },
        "required": [
          "loan_id",
          "borrower",
          "locality",
          "tranche_number",
          "milestone",
          "status",
          "request_amount",
          "recommendation",
          "recommendation_tone",
          "exposure",
          "exposure_undefined",
          "needs_human_review",
          "confidence",
          "stages",
          "math",
          "photos",
          "owner_view",
          "officer_view",
          "phases"
        ],
        "title": "TrancheDecisionView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Serves pipeline-shaped loan data to the Neev owner and bank consoles.",
    "title": "Neev API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/loans/{loan_id}": {
      "get": {
        "operationId": "loan_summary_api_loans__loan_id__get",
        "parameters": [
          {
            "description": "The loan's id, e.g. 1001.",
            "in": "path",
            "name": "loan_id",
            "required": true,
            "schema": {
              "description": "The loan's id, e.g. 1001.",
              "title": "Loan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanSummaryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Loan Summary",
        "tags": [
          "loans"
        ]
      }
    },
...
    "/api/loans/{loan_id}/milestones": {
      "post": {
        "description": "A borrower reporting progress: a stage, some photos, an optional note.\n\nThe photos attach to the tranche currently under review, which is the one\nwhose release they are evidence for.\n\nEvery field this endpoint cannot actually establish is left null: the\ngeotag, timestamp and same-angle flags, which the visual inspector fills in\nlive mode, and `taken_at` \u2014 upload time is not capture time, and writing it\ninto an evidence trail as though it were would be the one lie a verification\nrecord cannot afford.",
        "operationId": "report_milestone_api_loans__loan_id__milestones_post",
        "parameters": [
          {
            "description": "The loan's id, e.g. 1001.",
            "in": "path",
            "name": "loan_id",
            "required": true,
            "schema": {
              "description": "The loan's id, e.g. 1001.",
              "title": "Loan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_report_milestone_api_loans__loan_id__milestones_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report Milestone",
        "tags": [
          "loans"
        ]
      }
    },
    "/api/loans/{loan_id}/progress": {
      "get": {
        "operationId": "build_progress_api_loans__loan_id__progress_get",
        "parameters": [
          {
            "description": "The loan's id, e.g. 1001.",
            "in": "path",
            "name": "loan_id",
            "required": true,
            "schema": {
              "description": "The loan's id, e.g. 1001.",
              "title": "Loan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildProgressView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Progress",
        "tags": [
          "loans"
        ]
      }
    },
    "/api/loans/{loan_id}/questions/send": {
      "post": {
        "description": "Marks the drafted questions sent.\n\nIdempotent: a second click does not re-stamp `sent_at` on a question already\nsent, and the count returned is how many questions are now with the\ncontractor \u2014 the number the screen says it sent, both times.",
        "operationId": "send_questions_api_loans__loan_id__questions_send_post",
        "parameters": [
          {
            "description": "The loan's id, e.g. 1001.",
            "in": "path",
            "name": "loan_id",
            "required": true,
            "schema": {
              "description": "The loan's id, e.g. 1001.",
              "title": "Loan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionsSentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Questions",
        "tags": [
          "loans"
        ]
      }
    },
...
    "/api/loans/{loan_id}/tranches/{tranche_number}": {
      "get": {
        "description": "The officer's view of one tranche. Bank-only: it carries the officer's\nnarrative alongside the borrower's.",
        "operationId": "tranche_decision_api_loans__loan_id__tranches__tranche_number__get",
        "parameters": [
          {
            "description": "The loan's id, e.g. 1001.",
            "in": "path",
            "name": "loan_id",
            "required": true,
            "schema": {
              "description": "The loan's id, e.g. 1001.",
              "title": "Loan Id",
              "type": "string"
            }
          },
          {
            "description": "1-based tranche number.",
            "in": "path",
            "name": "tranche_number",
            "required": true,
            "schema": {
              "description": "1-based tranche number.",
              "minimum": 1,
              "title": "Tranche Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrancheDecisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tranche Decision",
        "tags": [
          "tranches"
        ]
      }
    },
    "/api/loans/{loan_id}/tranches/{tranche_number}/decision": {
      "post": {
        "description": "Append an audit event and update the latest-state projection atomically.\n\nClients can deduplicate retries using Idempotency-Key. Reusing a key with a\ndifferent payload is a conflict; an unkeyed submission is a new decision.",
        "operationId": "decide_api_loans__loan_id__tranches__tranche_number__decision_post",
        "parameters": [
          {
            "description": "The loan's id, e.g. 1001.",
            "in": "path",
            "name": "loan_id",
            "required": true,
            "schema": {
              "description": "The loan's id, e.g. 1001.",
              "title": "Loan Id",
              "type": "string"
            }
          },
          {
            "description": "1-based tranche number.",
            "in": "path",
            "name": "tranche_number",
            "required": true,
            "schema": {
              "description": "1-based tranche number.",
              "minimum": 1,
              "title": "Tranche Number",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide",
        "tags": [
          "tranches"
        ]
      }
    },
...
  }
}